requirements.txt
# Web Framework
flask==2.0.1
werkzeug==2.0.1
gunicorn==20.1.0
jinja2==3.0.1

# Data Processing
numpy==1.22.4
pandas==1.4.2
scikit-learn==1.0.2

# Deep Learning - Using CPU-only version to avoid compatibility issues
tensorflow-cpu==2.8.0

# Natural Language Processing
nltk==3.7

# Other utilities
click==8.0.1
markupsafe==2.0.1
